# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evatrack API",
        "description": "API Documentation for Evatrack Evacuation Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/households/{householdId}/members": {
            "get": {
                "tags": [
                    "Households"
                ],
                "summary": "List household members",
                "operationId": "e62afdf49059eebc2868d8a5f8820428",
                "parameters": [
                    {
                        "name": "householdId",
                        "in": "path",
                        "description": "Household ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Household not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Households"
                ],
                "summary": "Add member to household",
                "operationId": "e4035dd264fb9a1fa531b7e3c6682a8e",
                "parameters": [
                    {
                        "name": "householdId",
                        "in": "path",
                        "description": "Household ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "birth_date"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "middle_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "relationship_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "civil_status_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "vulnerable_group_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Household not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/households/{householdId}/members/{memberId}": {
            "delete": {
                "tags": [
                    "Households"
                ],
                "summary": "Delete household member",
                "operationId": "27b7b177a66a069403fa7008c6059fe6",
                "parameters": [
                    {
                        "name": "householdId",
                        "in": "path",
                        "description": "Household ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "memberId",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "400": {
                        "description": "Member is currently evacuated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Member or Household not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Households"
                ],
                "summary": "Update household member",
                "operationId": "437e6b20a9b47bcdeaeb77b643efedc7",
                "parameters": [
                    {
                        "name": "householdId",
                        "in": "path",
                        "description": "Household ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "memberId",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "birth_date"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "middle_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "relationship_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "civil_status_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "vulnerable_group_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Member or Household not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}